+++ BE/src/main/java/nem/com/service/impl/OrderServiceOnlineImpl.java
@@ -45,7 +45,7 @@ public class OrderServiceOnlineImpl implements OrderServiceOnline {
         return this.ordersRepository.getAllOrders(id) ;
     }
     @Override
-    public void updateStatusOrder(Short status, Long id) {
+    public void updateStatusOrder(Integer status, Long id) {
         List<OrderDetails> listOrderDetail = this.orderDetailsRepository.getOrderDetailsByOrder(id);
         for(OrderDetails x : listOrderDetail ){
             Integer a = x.getQuantity() + x.getProductsDetail().getQuantity() ;
@@ -53,6 +53,4 @@ public class OrderServiceOnlineImpl implements OrderServiceOnline {
         }
         this.ordersRepository.updateStatusOrder( status , id );
     }
-
-
 }
